create layout fro post
