+++ templates/post/post_video.php
@@ -16,7 +16,8 @@
 <div class="post-video__block">
     <div class="post-video__preview">
         <?= embed_youtube_cover(htmlspecialchars($post["contain"])); ?>
-        <img src="img/coast-medium.jpg" alt="Превью к видео <?= htmlspecialchars($post["title"]) ?>" width="360" height="188">
+        <img src="img/coast-medium.jpg" alt="Превью к видео <?= htmlspecialchars($post["title"]) ?>" width="360"
+             height="188">
     </div>
     <a href="post-details.html" class="post-video__play-big button">
         <svg class="post-video__play-big-icon" width="14" height="14">
